+++ Frends.Redis.SetValue/Frends.Redis.SetValue/Definitions/Input.cs
@@ -1,4 +1,7 @@
-﻿namespace Frends.Redis.SetValue.Definitions;
+﻿using System.Collections.Generic;
+using System.ComponentModel;
+
+namespace Frends.Redis.SetValue.Definitions;
 
 using System.ComponentModel.DataAnnotations;
 
@@ -8,6 +11,13 @@ using System.ComponentModel.DataAnnotations;
 public class Input
 {
     /// <summary>
+    /// Type of the value we want to save in Redis
+    /// </summary>
+    /// <example>ValueType.String</example>
+    [DefaultValue(ValueType.String)]
+    public ValueType ValueType { get; set; } = ValueType.String;
+
+    /// <summary>
     /// Key to set
     /// </summary>
     /// <example>Foo</example>
@@ -15,11 +25,26 @@ public class Input
     public string Key { get; set; }
 
     /// <summary>
-    /// Value to set
+    /// String value to set
     /// </summary>
     /// <example>Bar</example>
     [DisplayFormat(DataFormatString = "Text")]
-    public string Value { get; set; }
+    [UIHint(nameof(ValueType), "", ValueType.String, ValueType.Json)]
+    public string StringValue { get; set; }
+
+    /// <summary>
+    /// List value to set
+    /// </summary>
+    /// <example>["Foo", "Bar"]</example>
+    [UIHint(nameof(ValueType), "", ValueType.List, ValueType.Set)]
+    public List<string> ListValue { get; set; }
+
+    /// <summary>
+    /// Dictionary value to set
+    /// </summary>
+    /// <example>{ { "Foo", "Bar" }, { "Moo", "Baz" } }</example>
+    [UIHint(nameof(ValueType), "", ValueType.Hash)]
+    public Dictionary<string, string> HashValue { get; set; }
 
     /// <summary>
     /// Time to live in seconds. Leave empty for infinite.