+++ Frends.Redis.SetValue/CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## [2.0.0] - 2026-03-03
+### Added
+- Added support for more Redis value types (string, hash, list, JSON, set)
+- **[Breaking Change]** Rename Input.Value to Input.StringValue to distinguish between values of different types.
+- Added support for new connection methods
+- 
 ## [1.1.0] - 2025-07-31
 ### Changed
-- Make default error message empty by default
+- Make the default error message empty by default
 
 ## [1.0.0] - 2025-05-13
 ### Changed